+++ src/FactsFactoryInterface.php
@@ -2,7 +2,7 @@
 namespace BOTK;
 
 Interface FactsFactoryInterface {
-	public function factualize($rawData, &$globalStorage=NULL);
+	public function factualize($rawData, object $globalStorage=null);
 	public function generateLinkedDataHeader();
 	public function generateLinkedDataFooter();
 	public function addToCounter($counter,$val=1);